requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
asciimatics==1.14.0
asgiref==3.6.0
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
autopage==0.5.1
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.0
click==8.1.3
cliff==4.1.0
cmaes==0.9.0
cmd2==2.4.2
colorama==0.4.5
colorlog==6.7.0
constantly==15.1.0
contourpy==1.0.6
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
dateparser==1.1.4
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.6
Django==4.2
django-cors-headers==3.14.0
dnspython==2.3.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.0.0
filelock==3.12.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-commas==2.1.0
flake8-isort==5.0.3
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
greenlet==2.0.1
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.13.0
incremental==22.10.0
iniconfig==1.1.1
ipykernel==6.15.2
ipython==8.4.0
ipywidgets==8.0.2
isort==5.11.2
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.1
jmespath==1.0.1
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.3
mypy==0.991
mypy-extensions==0.4.3
mysql-connector-python==8.0.32
nest-asyncio==1.5.5
nltk==3.7
numpy==1.24.1
optuna==3.0.4
outcome==1.2.0
packaging==23.1
pandas==1.5.2
parsel==1.8.1
parso==0.8.3
pbr==5.11.0
pep8-naming==0.13.2
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.2.0
pluggy==1.0.0
prettytable==3.5.0
prompt-toolkit==3.0.30
Protego==0.2.1
protobuf==3.20.3
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pycparser==2.21
PyDispatcher==2.0.7
pyfiglet==0.8.post1
pyflakes==2.5.0
Pygments==2.13.0
pymongo==4.3.3
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
PySocks==1.7.1
pytest==7.2.0
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==6.0
pyzmq==23.2.1
queuelib==1.6.2
regex==2022.10.31
requests==2.30.0
requests-file==1.5.1
scikit-learn==1.1.3
scipy==1.8.1
Scrapy==2.9.0
seaborn==0.12.1
selenium==4.9.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
snscrape==0.4.3.20220106
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.3
stack-data==0.5.0
stevedore==4.1.1
TA-Lib @ file:///C:/Users/matro/TA_Lib-0.4.24-cp39-cp39-win_amd64.whl
tabulate==0.9.0
threadpoolctl==3.1.0
tldextract==3.4.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ujson==5.6.0
urllib3==1.26.15
virtualenv==20.21.0
w3lib==2.1.1
wcwidth==0.2.5
webcolors==1.13
websockets==10.4
widgetsnbextension==4.0.3
wsproto==1.2.0
yarl==1.8.2
zipp==3.11.0
zope.interface==6.0
